fix(stella-core): split the forked-skill scoping tests out of subagent/tests.rs
The wave-1 integration appended the #2682 test block to a file that was
already near the 1500-line ratchet, pushing it to 1567; the block moves to
its own sibling module borrowing the shared scaffolding, and fleet_cmd.rs
returns to its recorded ceiling by merging three crate imports into one line.

Refs #2682
